Alter css makes pages and edit app.jsx
