+++ desktop-web/src/components/Footer.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function Footer() {
+  return (
+    <div className='footer d-flex flex-column flex-lg-row bg-dark text-light p-2 p-lg-4'>
+      <div className=''>1</div>
+      <div>2</div>
+      <div>3</div>
+      <div>4</div>
+    </div>
+  )
+}
