+++ desktop-web/src/pages/HomePage.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export default function HomePage() {
+  return (
+   <div className='main-container'>
+
+      <h1 className="fs-4 my-4 text-center">Home page</h1>
+
+        {/* <div className='d-flex flex-column gap-4'>
+
+        </div> */}
+
+    </div>
+  )
+}
